Autocomplete: Refactor completion sources
