+++ editor/src/kroqed-editor/codeview/autocomplete/coqWords.json
