+++ editor/src/kroqed-editor/codeview/autocomplete/emojis.json
