+++ editor/src/kroqed-editor/codeview/autocomplete/emojis.ts
@@ -1,0 +1,23 @@
+import { Completion, CompletionContext, CompletionSource } from "@codemirror/autocomplete";
+import emojis from "./emojis.json";
+
+const emojiCompletions: Completion[] = emojis;
+
+/**
+ * Function that creates the `emojiCompletionSource`.
+ * This function can be used in the editor as a completion source. 
+ * 
+ * Emoji list has been generated by using https://github.com/github/gemoji/blob/master/db/emoji.json
+ * from https://github.com/github/gemoji 
+ */
+export const emojiCompletionSource: CompletionSource = function(context: CompletionContext) {
+    let before = context.matchBefore(/:.*/);    
+    // If completion wasn't explicitly started and there
+    // is no word before the cursor, don't open completions.
+    if (!context.explicit && !before) return null;
+    return {
+        from: before ? before.from : context.pos,
+        options: emojiCompletions,
+        validFor: /^.*$/
+    }
+}
