+++ editor/src/kroqed-editor/codeview/autocomplete/index.ts
@@ -1,0 +1,4 @@
+// Exports for this class
+export { emojiCompletionSource } from "./emojis";
+export { coqCompletionSource } from "./coqTerms";
+export { symbolCompletionSource } from "./symbols";
