+++ editor/src/kroqed-editor/codeview/autocomplete/symbols.json
