+++ editor/src/kroqed-editor/codeview/lang-pack/index.ts
@@ -8,7 +8,6 @@ import { Tag, styleTags, tags as t } from "@lezer/highlight"
 
 // Importing the parser for the Coq language
 import { parser } from "./syntax"
-import coqWords from "./coqWords.json"
 
 // Defining custom tags for specific elements of the Coq language
 let customTags = {
@@ -48,22 +47,6 @@ export const coqLanguage = LRLanguage.define({
         commentTokens: { line: "(*" }
     }
 })
-
-// Our list of completions (can be static, since the editor
-/// will do filtering based on context).
-const coqCompletions: Completion[] = coqWords;
-
-export const coqCompletionSource: CompletionSource = function(context) {
-    let before = context.matchBefore(/(\w+\-*\s*)+/)
-    // If completion wasn't explicitly started and there
-    // is no word before the cursor, don't open completions.
-    if (!context.explicit && !before) return null
-    return {
-        from: before ? before.from : context.pos,
-        options: coqCompletions,
-        validFor: /^\w*$/
-    }
-}
 
 export function coq() {
     return new LanguageSupport(coqLanguage)
