--- editor/src/kroqed-editor/completions/index.ts
@@ -1,2 +1,0 @@
-// Exports for this class
-export { symbolCompletionSource, emojiCompletionSource } from "./completions";
